Fix: (refs #183) time_management> avoid '<=' to remove conflict at the edge of range
If I do `ends[i] <= time <= ends[i+1]`,
This expression will be True twice when `time` is the same as
`ends[i+1]` because that would be the same as `ends[i] for the
next expression.

Of cource, this conflict won't work because of looping,
But it is not good idea to leave things confusing, right?
So I fixed it