Change: api/swagger> specify Loader of yaml.load() to suspend warning ...
The default Loader(UnsafeLoader) is not recommended so that we should
specify the Loader for safety.
In this case, as we don't treat any untrusted yaml file('cuz what we
load is one of our file), we don't have to worry about injection.
So I select 'FullLoader', which supports full yaml language and suspend
arbitrary code execution.

See more details at:
  https://github.com/yaml/pyyaml/wiki/PyYAML-yaml.load(input)-Deprecation)